requirements.txt
python-dotenv
requests
pyarrow
pandas